move pretix types to pcd/passport-interface
